requirements.txt
Django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
psycopg2-binary
pytest
requests
pytest-django